DBZエラー除去

+++ DB/DBZ/standard/business/src/main/java/jp/co/ndensan/reams/db/dbz/business/core/basic/YoboKeikakuJikoSakuseiMeisai.java
@@ -44,7 +44,7 @@ public class YoboKeikakuJikoSakuseiMeisai extends
      */
     public YoboKeikakuJikoSakuseiMeisai(HihokenshaNo 被保険者番号,
             FlexibleYearMonth 対象年月,
-            Decimal 履歴番号,
+            Integer 履歴番号,
             RString 居宅サービス区分,
             JigyoshaNo サービス提供事業者番号,
             ServiceShuruiCode サービス種類コード,
@@ -131,7 +131,7 @@ public class YoboKeikakuJikoSakuseiMeisai extends
      *
      * @return 履歴番号
      */
-    public Decimal get履歴番号() {
+    public Integer get履歴番号() {
         return entity.getRirekiNo();
     }
