+++ DB/DBZ/standard/business/src/test/java/jp/co/ndensan/reams/db/dbz/business/core/basic/KyotakuKeikakuJigyoshaSakuseiIdentifierTest.java
@@ -28,7 +28,7 @@ public class KyotakuKeikakuJigyoshaSakuseiIdentifierTest extends DbzTestBase {
 //TODO 主キーの数が足りない場合、追加してください。
     private static HihokenshaNo 被保険者番号;
     private static FlexibleYearMonth 対象年月;
-    private static Decimal 履歴番号;
+    private static Integer 履歴番号;
 
     @BeforeClass
     public static void setUpClass() {
