+++ DB/DBZ/standard/persistence/src/main/java/jp/co/ndensan/reams/db/dbz/persistence/db/basic/DbT3006KyotakuKeikakuJigyoshaSakuseiDac.java
@@ -43,7 +43,7 @@ public class DbT3006KyotakuKeikakuJigyoshaSakuseiDac implements ISaveable<DbT300
     @Transaction
     public DbT3006KyotakuKeikakuJigyoshaSakuseiEntity selectByKey(HihokenshaNo 被保険者番号,
             FlexibleYearMonth 対象年月,
-            Decimal 履歴番号) throws NullPointerException {
+            Integer 履歴番号) throws NullPointerException {
         requireNonNull(対象年月, UrSystemErrorMessages.値がnull.getReplacedMessage("対象年月"));
         requireNonNull(履歴番号, UrSystemErrorMessages.値がnull.getReplacedMessage("履歴番号"));
